requirements.txt
openai~=1.6.1
google-generativeai~=0.5.3
pandas~=2.2.1
swiplserver~=1.0.2
tiktoken~=0.6.0